refactor(watcher): extract _WatchedFieldBase to consolidate sync and async field logic
The two WatchedField classes shared ~90% of their implementation but had
no common base, causing drift (the sync version acquired a lock in _update
but then read _value outside it for the callback; the async version had no
lock at all). Extract _WatchedFieldBase with _apply_raw and _fire_callbacks
helpers. Both subclasses call super().__init__, inherit path/on_change/__bool__,
and delegate value-mutation to the base—with or without a lock as appropriate.
The three _RECONNECT_* constants are also consolidated in the new module.

Closes #60
